Fix fetch request tests
